refactor: migrate constants (plans) from apps/web to @core/components
