+++ apps/web/ce/components/workspace/billing/comparison/root.tsx
@@ -9,8 +9,8 @@ import { observer } from "mobx-react";
 import type { EProductSubscriptionEnum, TBillingFrequency } from "@plane/types";
 // components
 import { PlansComparisonBase, shouldRenderPlanDetail } from "@/components/workspace/billing/comparison/base";
-import type { TPlanePlans } from "@/constants/plans";
-import { PLANE_PLANS } from "@/constants/plans";
+import type { TPlanePlans } from "@/components/workspace/billing/comparison/plans";
+import { PLANE_PLANS } from "@/components/workspace/billing/comparison/plans";
 // plane web imports
 import { PlanDetail } from "./plan-detail";
