+++ apps/web/core/components/workspace/billing/comparison/index.ts
@@ -5,3 +5,4 @@
  */
 
 export * from "./base";
+export * from "./plans";
